requirements.txt
customtkinter==5.2.2
pillow==10.2.0
tk==0.1.0
torch==2.2.1
torchvision==0.17.1
